add api folder and mocks; add app js file
