+++ api/routes.js
@@ -1,0 +1,5 @@
+export default {
+    laws: '/laws',
+    law: lawId => `/laws/${lawId}`,
+    lawTypes: '/law_types',
+}
